requirements.txt
Flask>=2.0
SQLAlchemy>=1.4
psycopg2-binary>=2.9
python-dotenv>=0.21
Werkzeug>=2.0
gunicorn>=20.1.0